requirements.txt
python-dotenv==1.0.1
requests==2.31.0
python-docx==1.1.0
pdfplumber==0.10.3 